Integrate ahrs communication and parsing logic
I added another device type, 'pc', which expects to communicate with the
ahrs on '/dev/ttyUSB0' and read the depth sensor from a file called
'depth_in' in the root directory of the repo.

The pc binary seems to work to some degree, but the arduino code gives
NaN for the motor powers right now. I'll try to fix that soon.

There is definitely some glue code going on, and naming is
somewhat...inconsistent.

I also made eye contact with Brian Kim at UCSD yesterday, lol