Enable global interrupts in avr io_ahrs_init
I assumed it wouldn't be necessary because the arduino libs would just
leave interrupts enabled globally all the time, but that doesn't seem to
be the case.

The NaN issue on the avr was because the config values hadn't actually
been put into EEPROM. The avr bootloader won't actually write EEPROM,
but I've done it with an isp programmer, and the setConfig function can
be used when it's implemented.