Switch to compiled-in config
This replaces config in eeprom on the arduino and in a file for cpu and
pc. This is because changing eeprom can be a pain on the arduino (the
'wiring' programmer protocol has no capability to do it directly), and
right now it is easier to just recompile and reflash.