Make linear movement absolute
Both linearly and rotationally.

Doing this has made be realize that rotational thrust is calculated only
relatively, despite the fact that measured pitch is absolute relative to
roll, and measured yaw is absolute relative to both pitch and roll, so
thrust applied will begin to disagree with sensor measuremens under
tilted conditions.